Fixed empty cachepwd field in activation process (again, ... sorry!)
The problem is, the original: $this->user->set('cachepwd',''); doesn't work as this generates a hash of an empty value and the cachepwd field isn't emptied.

This is a problem in the MODX core and needs to be fixed.

My first fix (previous PR), using the _setRaw method also doesn't work in some cases, as this method is protected.

Signed-off-by: gadgetto <rij-Ub-eiD-yab-Yak-Ek>